Hide the expression implementation behind an interface
