+++ jmespath-core/src/main/java/io/burt/jmespath/JmesPathRuntime.java
@@ -16,6 +16,8 @@ public interface JmesPathRuntime<T> extends Comparator<T> {
   /**
    * Compile a JMESPath expression into a reusable expression object.
    *
+   * The expression objects should be stateless and thread safe.
+   *
    * @throw ParseException when the string is not a valid JMESPath expression
    */
   public JmesPathExpression<T> compile(String expression);
