+++ jmespath-core/src/main/java/io/burt/jmespath/parser/JmesPathExpressionParser.java
@@ -11,6 +11,7 @@ import org.antlr.v4.runtime.tree.ParseTree;
 
 import io.burt.jmespath.JmesPathExpression;
 import io.burt.jmespath.JmesPathRuntime;
+import io.burt.jmespath.StandardExpression;
 import io.burt.jmespath.node.AndNode;
 import io.burt.jmespath.node.ComparisonNode;
 import io.burt.jmespath.node.CreateArrayNode;
@@ -77,7 +78,7 @@ public class JmesPathExpressionParser<T> extends JmesPathBaseVisitor<JmesPathNod
   }
 
   public JmesPathExpression<T> expression() {
-    return new JmesPathExpression<T>(runtime, visit(tree));
+    return new StandardExpression<T>(runtime, visit(tree));
   }
 
   private String identifierToString(JmesPathParser.IdentifierContext ctx) {
